Refactor reusable functions out of TargetCli class
Functional changes:
* Target-shell ls supports -h flag
* Target-fs ls now outputs with color when printing to terminal
