+++ tests/tools/test_shell.py
@@ -3,14 +3,11 @@ import platform
 import sys
 from io import BytesIO, StringIO
 from pathlib import Path
-from unittest.mock import MagicMock
 
 import pytest
 
-from dissect.target.exceptions import FileNotFoundError
-from dissect.target.filesystem import FilesystemEntry
-from dissect.target.helpers.fsutil import TargetPath, normalize, stat_result
-from dissect.target.tools import shell
+from dissect.target.helpers.fsutil import normalize
+from dissect.target.tools import fsutils
 from dissect.target.tools.shell import (
     TargetCli,
     TargetHubCli,
@@ -101,7 +98,7 @@ def test_targethubcli_autocomplete_enter(make_mock_targets):
     assert suggestions == ["1"]
 
 
-def test_targetcli_autocomplete(target_bare):
+def test_targetcli_autocomplete(target_bare, monkeypatch):
     target_cli = TargetCli(target_bare)
 
     base_path = "/base-path/"
@@ -115,7 +112,7 @@ def test_targetcli_autocomplete(target_bare):
             (None, subpath_mismatch),
         ]
 
-    target_cli.scandir = dummy_scandir
+    monkeypatch.setattr("dissect.target.tools.shell.ls_scandir", dummy_scandir)
 
     suggestions = target_cli.completedefault("sub", f"ls {base_path}sub", 3 + len(base_path), 3 + len(base_path) + 3)
     assert suggestions == [subpath_match]
@@ -163,61 +160,13 @@ def test_exec_target_command(capfd, target_default):
 
 def test_target_cli_ls(target_win, capsys, monkeypatch):
     # disable colorful output in `target-shell`
-    monkeypatch.setattr(shell, "LS_COLORS", {})
+    monkeypatch.setattr(fsutils, "LS_COLORS", {})
 
     cli = TargetCli(target_win)
     cli.onecmd("ls")
 
     captured = capsys.readouterr()
     assert captured.out == "\n".join(["c:", "sysvol"]) + "\n"
-
-
-def test_target_cli_print_extensive_file_stat(target_win, capsys):
-    mock_stat = stat_result([0o1777, 1, 2, 3, 1337, 7331, 999, 0, 0, 0])
-    mock_entry = MagicMock(spec_set=FilesystemEntry)
-    mock_entry.lstat.return_value = mock_stat
-    mock_entry.is_symlink.return_value = False
-    mock_path = MagicMock(spec_set=TargetPath)
-    mock_path.get.return_value = mock_entry
-
-    mock_args = MagicMock()
-
-    cli = TargetCli(target_win)
-    cli.print_extensive_file_stat(mock_args, sys.stdout, mock_path, "foo")
-
-    captured = capsys.readouterr()
-    assert captured.out == "-rwxrwxrwx 1337 7331    999 1970-01-01T00:00:00 foo\n"
-
-
-def test_target_cli_print_extensive_file_stat_symlink(target_win, capsys):
-    mock_stat = stat_result([0o1777, 1, 2, 3, 1337, 7331, 999, 0, 0, 0])
-    mock_entry = MagicMock(spec_set=FilesystemEntry)
-    mock_entry.lstat.return_value = mock_stat
-    mock_entry.is_symlink.return_value = True
-    mock_entry.readlink.return_value = "bar"
-    mock_path = MagicMock(spec_set=TargetPath)
-    mock_path.get.return_value = mock_entry
-
-    mock_args = MagicMock()
-
-    cli = TargetCli(target_win)
-    cli.print_extensive_file_stat(mock_args, sys.stdout, mock_path, "foo")
-
-    captured = capsys.readouterr()
-    assert captured.out == "-rwxrwxrwx 1337 7331    999 1970-01-01T00:00:00 foo -> bar\n"
-
-
-def test_target_cli_print_extensive_file_stat_fail(target_win, capsys):
-    mock_path = MagicMock(spec_set=TargetPath)
-    mock_path.get.side_effect = FileNotFoundError("ERROR")
-
-    mock_args = MagicMock()
-
-    cli = TargetCli(target_win)
-    cli.print_extensive_file_stat(mock_args, sys.stdout, mock_path, "foo")
-
-    captured = capsys.readouterr()
-    assert captured.out == "??????????    ?    ?      ? ????-??-??T??:??:??.?????? foo\n"
 
 
 @pytest.mark.parametrize(
